requirements.txt
alembic==0.9.9
amqp==2.6.0
apache-airflow==1.10.10
apispec==1.3.3
argcomplete==1.12.0
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
blinker==1.4
bs4==0.0.1
cached-property==1.5.1
cattrs==0.9.2
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.34
cryptography==2.9.2
defusedxml==0.6.0
dill==0.3.2
dnspython==1.16.0
docutils==0.16
email-validator==1.1.1
Flask==1.0.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-Dance==0.14.0
Flask-DebugToolbar==0.10.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-OAuth==0.12
Flask-OAuthlib==0.9.4
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.13
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
flower==0.9.5
funcsigs==1.0.2
future==0.18.2
graphviz==0.14.1
gunicorn==19.10.0
httplib2==0.11.3
humanize==2.5.0
idna==2.6
importlib-metadata==1.7.0
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
json-merge-patch==0.2
jsonschema==3.2.0
kafka-python==2.0.1
kombu==4.6.11
lazy==1.3
lazy-object-proxy==1.5.0
lockfile==0.12.2
lxml==4.5.2
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.23.1
mysqlclient==1.3.14
natsort==7.0.1
numpy==1.19.0
oauth2==1.9.0.post1
oauthlib==2.0.7
pandas==0.25.3
pendulum==1.4.4
pip==20.0.2
prison==0.1.3
prometheus-client==0.8.0
psutil==5.7.0
psycopg2-binary==2.8.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.11.0
pyrsistent==0.16.0
python-daemon==2.1.1
python-dateutil==2.7.2
python-editor==1.0.3
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.18.4
requests-oauthlib==0.8.0
rq==1.5.1
setproctitle==1.1.10
setuptools==45.3.0
six==1.11.0
soupsieve==2.0.1
SQLAlchemy==1.2.6
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.33.2
tabulate==0.8.7
tenacity==4.12.0
termcolor==1.1.0
text-unidecode==1.2
thrift==0.13.0
tornado==5.1.1
typing-extensions==3.7.4.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.22
URLObject==2.4.3
vine==1.3.0
Werkzeug==0.14.1
wheel==0.35.1
wincertstore==0.2
WTForms==2.1
zipp==3.1.0
zope.deprecation==4.4.0